#6 validated the file using HTML validator
I've validated the file and cleared all the mistakes. Put all the content of the letter inside the body tag and html layout inside the head tag. Also, added alignment inside the <p> tag, instead of using <ul> for bullet points I've changed it to <ol>. and cleared the mistakes in the opening tags in the last lines of code.